services/renderer/pyproject.toml
[project]
name = "dadi-docx-renderer"
version = "0.1.0"
description = "DOCX-first deterministic rendering stage for DADI-style pipelines (Stage 06)."
requires-python = ">=3.10"
dependencies = [
  "psycopg[binary]>=3.1",
  "pydantic>=2.0",
  "python-docx>=1.1.0",
]
